% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanban_methods.R
\name{removeBoard}
\alias{removeBoard}
\alias{addBoard}
\alias{addElement}
\title{Methods to interact with your boards}
\usage{
removeBoard(inputId, boardId, session = shiny::getDefaultReactiveDomain())

addBoard(inputId, boardId, session = shiny::getDefaultReactiveDomain())

addElement(
  inputId,
  boardId,
  itemId,
  title,
  ...,
  position = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{shiny input}

\item{boardId}{Id of the board to interact with}

\item{session}{Shiny Session object}

\item{itemId}{Id of the item to add}

\item{title}{Content of the item to add}

\item{...}{further args passed to the JS object}

\item{position}{If position is set, inserts at position starting from 0}
}
\description{
Methods to interact with your boards
}
